+++ models/inc/move.h
@@ -53,7 +53,7 @@ typedef enum {
 typedef struct {
     move_model_type_t type;
     char* name;
-    int32_t hp;
+    int16_t hp;
     uint16_t weight;
     uint16_t speed;
 } move_model_t;
